requirements.txt
playwright>=1.48.0
beautifulsoup4>=4.12.0
azure-cosmos>=4.7.0
